* develop: (29 commits)
  build(deps-dev): Bump the npm-dependencies group in /WebUI with 11 updates (#127)
  Remove loop() behaviour from code (#123)
  Check if link code is proper length + use numeric input (#124)
  RGB Support plus OpenShock Core V1 board definition (#121)
  Increase stability by handling heavy tasks better (#122)
  D1 Mini default pin should be 15
  Rework HTTP/JSON handling to be cleaner (#106)
  Push color palette (#118)
  build(deps-dev): Bump the npm-dependencies group in /WebUI with 9 updates (#117)
  build(deps-dev): Bump the npm-dependencies group (#116)
  Decrease cache to 1 hour (#115)
  Fix captive portal not showing network savestate + forget functionality (#113)
  Improve git commit hash CPP define (#110)
  Conform EStopManager to code style (#109)
  More code cleanup (#107)
  Bugfix/shocker unpaired randomly (#105)
  Some more code cleanup (#104)
  Fix serial WiFi configuration not working (#102)
  Better captive portal logging (#103)
  Add http caching with filesystem hash to reduce load on esp (#101)
  ...

+++ src/CommandHandler.cpp
@@ -4,7 +4,7 @@
 #include "Config.h"
 #include "Constants.h"
 #include "Logging.h"
-#include "RFTransmitter.h"
+#include "radio/RFTransmitter.h"
 
 #include <memory>
 
@@ -15,7 +15,7 @@ using namespace OpenShock;
 static std::unique_ptr<RFTransmitter> s_rfTransmitter = nullptr;
 
 bool CommandHandler::Init() {
-  std::uint32_t txPin = Config::GetRFConfig().txPin;
+  std::uint8_t txPin = Config::GetRFConfig().txPin;
   if (!OpenShock::IsValidOutputPin(txPin)) {
     ESP_LOGW(TAG, "Clearing invalid RF TX pin");
     Config::SetRFConfigTxPin(Constants::GPIO_INVALID);
